requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy<2.0.0
openpyxl>=3.0.0
plotly>=5.0.0
odoorpc>=0.9.0
python-dotenv>=1.0.0
